requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-dotenv
google-generativeai
jinja2
python-multipart